+++ geotrek/api/v2/serializers.py
@@ -913,6 +913,7 @@ if 'geotrek.sensitivity' in settings.INSTALLED_APPS:
         geometry = geo_serializers.GeometryField(read_only=True, source="geom_transformed", precision=7)
         species_id = serializers.SerializerMethodField()
         kml_url = serializers.SerializerMethodField()
+        attachments = AttachmentSerializer(many=True)
 
         def get_name(self, obj):
             return get_translation_or_dict('name', self, obj.species)
@@ -941,7 +942,7 @@ if 'geotrek.sensitivity' in settings.INSTALLED_APPS:
                 'id', 'contact', 'description', 'elevation',
                 'geometry', 'info_url', 'kml_url', 'name', 'period',
                 'practices', 'published', 'species_id', 'provider', 'structure',
-                'url'
+                'url', 'attachments'
             )
 
     class BubbleSensitiveAreaSerializer(SensitiveAreaSerializer):