+++ geotrek/sensitivity/tests/test_functional.py
@@ -21,6 +21,7 @@ class SensitiveAreaViewsTests(GeotrekAPITestCase, CommonTest):
 
     def get_expected_json_attrs(self):
         return {
+            'attachments': [],
             'contact': '<a href="mailto:toto@tata.com">toto@tata.com</a>',
             'description': 'Blabla',
             'kml_url': '/api/en/sensitiveareas/{}.kml'.format(self.obj.pk),
@@ -91,6 +92,7 @@ class RegulatorySensitiveAreaViewsTests(GeotrekAPITestCase, CommonTest):
 
     def get_expected_json_attrs(self):
         return {
+            'attachments': [],
             'contact': '<a href="mailto:toto@tata.com">toto@tata.com</a>',
             'description': 'Blabla',
             'kml_url': '/api/en/sensitiveareas/{}.kml'.format(self.obj.pk),